[FIX] survey: only show attendees ranking when relevant
The computed field 'session_show_ranking' was only computed based on the fact
that the survey had a scoring mechanism.

It's not very accurate since we also need a question that saves as the attendee
nickname to be able to display a relevant ranking chart.

This commit fixes the compute method by adding the necessary condition.

Task ID 2188482

closes odoo/odoo#45852

X-original-commit: 0256f0ad9ad87d38649fa18703d0869382296d28
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>