[FIX] payment: prevent ACL denial when unlinking onboarding wizard
Since wizards shall not be unlinked by anyone, but as we want to get rid
of this specific one immediately to get rid of eventual credentials,
do it with elevated privileges.

Before this commit, the behavior was an authorization error when trying
to applying the payment method, in a fresh invoicing configuration.

closes odoo/odoo#45363

X-original-commit: 6eb38cd61b0a4da9ce8c6154b6b5ce6614a659bb
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>