[FIX] point_of_sale: long product name print receipt in safari
Fine-tunning of 263131f49828d2d8f2177bae15d0af84af341c37

Before this commit, when printing in safari a receipt that contains
products with long names, parts of the name were overlapped. This occurs
because, in safari, the library that transforms the html into text
mismanages the spaces.

Now, the name is printed correctly in safari, same behaviour than the
others browsers.

opw-2169656

closes odoo/odoo#45794

X-original-commit: 9e4ab0da0e84b2b792659477e3520f7847d51aac
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>