[REF] web_editor: remove useless code in palette
Previously the widget was not refreshed everytime it was opened.
Now that it is the case and since the color picker rerenders every time
we select a color:
It is no longer required to select/unselect the colors.
It is no longer required to add custom colors to existing colors.

Part of: https://github.com/odoo/odoo/pull/45872

task-2212288

closes odoo/odoo#47355

X-original-commit: 8b03e49091081e54ca96c7b77aaed99e11f54acb
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>